+++ R/sptable.r
@@ -15,7 +15,7 @@
 #' 
 #' @param x \code{\link[sp]{Spatial}} object
 #' @param ... ignored
-#' @return \code{\link[dplyr]{tbl_df}} data_frame with columns
+#' @return \code{\link[tibble]{tibble}}  with columns
 #' \itemize{
 #'  \item SpatialPolygonsDataFrame "object_"   "branch_"   "island_"   "order_" "x"    "y_"
 #'  \item SpatialLinesDataFrame "object_"   "branch_" "order_"  "x_"      "y_"
@@ -24,7 +24,7 @@
 #' }
 #' @export
 #' @examples 
-#' ## holey is decomposed SpatialPolygonsDataFrame
+#' ## holey is a decomposed SpatialPolygonsDataFrame
 #' spdata <- sp(holey)
 #' library(sp)
 #' plot(spdata, col = rainbow(nrow(spdata), alpha = 0.4))
@@ -45,11 +45,10 @@
 #' ## example using in-place modification with sptable<-
 #' library(maptools)
 #' data(wrld_simpl)
-#' holey2 <- holey
+#' spdata2 <- spdata
 #' library(dplyr)
-## modify the geometry on this object without separating the vertices from the objects
-#' sptable(holey2) <- sptable(woz) %>% mutate(y_ = y_ + 5)
-
+#' ## modify the geometry on this object without separating the vertices from the objects
+#' sptable(spdata2) <- sptable(spdata2) %>% mutate(x_ = x_ + 10, y_ = y_ + 5)
 sptable <- function(x, ...) {
   UseMethod("sptable")
 }
@@ -87,9 +86,9 @@ sptable.SpatialMultiPointsDataFrame <- function(x, ...) {
    df
 }
 ## TODO multipoints
-#' @importFrom dplyr data_frame
+#' @importFrom tibble as_tibble
 mat2d_f <- function(x) {
-  as_data_frame(as.data.frame((x)))
+  as_tibble(as.data.frame((x)))
 }
 
 
@@ -141,6 +140,7 @@ mat2d_f <- function(x) {
 ## generalized on Polygon and Line
 #' @importFrom sp geometry
 #' @importFrom dplyr bind_rows
+#' @importFrom tibble tibble
 .gobbleGeom <-   function(x,  ...) {
   gx <- geometry(x)
   typ <- switch(class(gx), 
@@ -161,11 +161,11 @@ mat2d_f <- function(x) {
                                  order_ = seq(nr),
                                  x_ = coords[,1], 
                                  y_ = coords[,2])
-                     as_data_frame(lst[!sapply(lst, is.null)])
+                     as_tibble(lst[!sapply(lst, is.null)])
                    }
       )
       psd <- do.call(bind_rows, ps)
-      objlist[[i]] <- bind_cols(data_frame(object_ = rep(i, nrow(psd))), psd)
+      objlist[[i]] <- bind_cols(tibble(object_ = rep(i, nrow(psd))), psd)
       cnt <- cnt + nsubobs
     }
   obs <- do.call(bind_rows, objlist)
@@ -181,16 +181,16 @@ mat2d_f <- function(x) {
 
 .pointsGeom <-  function(x, ...) {
   ## this will have to become a tbl
-  xy <- as_data_frame(as.data.frame(coordinates(x)))
+  xy <- as_tibble(as.data.frame(coordinates(x)))
   cnames <- c('object_', 'x_', 'y_')
   ##xy <- cbind(1:nrow(xy), xy)
   if (is.list(x@coords)) {
     br <- rep(seq_along(x@coords), unlist(lapply(x@coords, nrow)))
     cnames <- c('branch_', 'object_', 'x_', 'y_')
-    xy <- bind_cols(data_frame(br), data_frame(br), xy)
+    xy <- bind_cols(tibble(br), tibble(br), xy)
   } else {
     br <- seq(nrow(xy))
-    xy <- bind_cols(data_frame(br), xy)
+    xy <- bind_cols(tibble(br), xy)
   }
   
   colnames(xy) <- cnames